requirements.txt
streamlit==1.22.0
openai==1.3.0
python-dotenv==1.0.0
transformers>=4.39.0
torch==1.13.1
sentence-transformers>=2.2.2
networkx==2.8.4
scikit-learn==1.2.2
PyPDF2==3.0.1
python-docx==0.8.11
SpeechRecognition==3.10.0
gTTS==2.3.1
Pillow==9.5.0
langchain==0.2.14
langchain_huggingface==0.0.3
langchain_community==0.2.12
google-api-python-client
streamlit-lottie
requests
streamlit-chat 
beautifulsoup4
plotly
pandas
numpy
docx2txt
tenacity
pydub
scipy
docx
markdown
pygame
opencv-python
playsound
sounddevice
reportlab
xldr